Rename ExecutorGroup and move getMaxExtent to Executor
